refactor: server, client 컴포넌트 명확히 분리
