+++ apps/web/app/home/components/home-banner.tsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import Image from 'next/image';
+
+export default function HomeBanner() {
+  return (
+    <div className="flex w-full justify-center bg-gray-300">
+      <Image
+        src={
+          'https://media.istockphoto.com/id/1154370446/ko/%EC%82%AC%EC%A7%84/%ED%9D%B0%EC%83%89-%EB%B0%B0%EA%B2%BD%EC%97%90-%EA%B3%A0%EB%A6%BD-%EB%90%9C-%EB%B0%94%EC%9C%84-%EC%A0%9C%EC%8A%A4%EC%B2%98%EB%A5%BC-%EB%B3%B4%EC%97%AC%EC%A3%BC%EB%8A%94-%EB%85%B9%EC%83%89-%EC%84%A0%EA%B8%80%EB%9D%BC%EC%8A%A4%EC%97%90-%EC%9E%AC%EB%AF%B8-%EB%84%88%EA%B5%AC%EB%A6%AC.jpg?s=1024x1024&w=is&k=20&c=mopsJIVkM2O1h3_jVXT6HErRa4coSU4g31IDbwDv2H4='
+        }
+        alt="배너"
+        width={1366}
+        height={560}
+        className="h-[56rem] w-[136.6rem] object-contain"
+      />
+    </div>
+  );
+}
